rock: update kernel to 5.10.218-Dv-Edition

Date:      Thu Dec 5 03:13:25 2024 +0700

+++ kernel-headers/usr/include/asm/sigcontext.h
@@ -74,7 +74,7 @@ struct fpsimd_context {
 	struct _aarch64_ctx head;
 	__u32 fpsr;
 	__u32 fpcr;
-	__u64 vregs[32];
+	__uint128_t vregs[32];
 };
 
 /*
